Add unit tests for the module system dialog
